added ProofArea,  Proof and Proofs components with test displaying proofs in it
