+++ src/components/Proofs/components/ProofsArea/Proof/Proof.jsx
@@ -1,0 +1,24 @@
+import { Box, Divider, ListItem, Typography } from "@mui/material";
+import { getRelativeTime } from "../../../../../shared/functions/getRelativeTime";
+
+const Proof = ({ date, title, description, author }) => {
+    
+    return (
+        <>
+            <ListItem>
+                <Box>
+					<Typography variant="h6" sx={{ fontWeight: 700, fontSize: 24 }}>
+						{title}
+						<Typography sx={{ fontSize: "10px", color: "#888888" }}>
+							{getRelativeTime(+new Date(date))}
+						</Typography>
+						<Typography sx={{ fontSize: "16px" }}>{description}</Typography>
+					</Typography>
+				</Box>
+            </ListItem>
+            <Divider variant="middle" component="li" sx={{marginBottom: 2}}/>
+        </>
+    )
+}
+
+export {Proof};
